Bug修复

+++ src/main/java/com/chongdao/client/repository/CartRepository.java
@@ -4,5 +4,5 @@ import com.chongdao.client.entitys.Carts;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CartRepository extends JpaRepository<Carts,Integer> {
-    int countByUserIdAndShopIdAndChecked(Integer userId,Integer shopId,Integer checked);
+    int countByUserIdAndShopIdAndChecked(Integer userId,Integer shopId,Byte checked);
 }
